# wso2-am-gateway/app/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Orders API",
    "description": "Sample Orders API for WSO2 APIM Gateway demo",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://app:3000",
      "description": "Backend service"
    }
  ],
  "paths": {
    "/api/orders": {
      "get": {
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "product": {
            "type": "string",
            "example": "Laptop"
          },
          "amount": {
            "type": "number",
            "example": 1200
          }
        }
      },
      "OrderInput": {
        "type": "object",
        "required": ["product", "amount"],
        "properties": {
          "product": {
            "type": "string",
            "example": "Monitor"
          },
          "amount": {
            "type": "number",
            "example": 300
          }
        }
      }
    }
  }
}